feat: add Android module, eBPF/kernel hooks, disassembler and agent crates
- src/android: dex/logcat/process support
- src/hook/ebpf_hook.rs, hw_breakpoint.rs, kernel_channel.rs
- src/disasm: disassembler module
- crates/frida-agent, crates/simple-injector workspace members
- examples/hook_wechat_input.rhai

+++ src/anti_detect/stack_fake.rs
@@ -373,13 +373,14 @@ impl StackWalker {
 }
 
 #[cfg(windows)]
+#[link(name = "dbghelp")]
 extern "system" {
     fn CaptureStackBackTrace(
         FramesToSkip: u32,
         FramesToCapture: u32,
         BackTrace: *mut *mut winapi::ctypes::c_void,
         BackTraceHash: *mut u32,
-    ) -> u16;
+    ) -> u32;
 }
 
 #[cfg(test)]
